Update bigint, zk-paillier verify and key_gen_second_message function calls

+++ src/ecdsa/two_party/mod.rs
@@ -15,6 +15,7 @@ use curv::cryptographic_primitives::hashing::hmac_sha512;
 use curv::cryptographic_primitives::hashing::traits::KeyedHash;
 use curv::elliptic::curves::traits::{ECPoint, ECScalar};
 use curv::{BigInt, FE, GE};
+use curv::arithmetic::big_num::{Pow, One};
 use multi_party_ecdsa::protocols::two_party_ecdsa::lindell_2017::{party_one, party_two};
 use paillier::*;
 
@@ -60,7 +61,7 @@ pub fn hd_key(
     pubkey: &GE,
     chain_code_bi: &BigInt,
 ) -> (GE, FE, GE) {
-    let mask = BigInt::from(2).pow(256) - BigInt::one();
+    let mask = BigInt::from(2).pow(256 as u32) - BigInt::one();
     // let public_key = self.public.q.clone();
 
     // calc first element:
